Work on GDT

+++ Kernel/clib/mem.h
@@ -8,9 +8,9 @@ void memcpy(char* src, char* dest, unsigned int num_bytes)
     }
 }
 
-void* memset(void* bufptr, int value, size_t size) {
+void* memset(void* bufptr, int value, int size) {
 	unsigned char* buf = (unsigned char*) bufptr;
-	for (size_t i = 0; i < size; i++)
+	for (int i = 0; i < size; i++)
 		buf[i] = (unsigned char) value;
 	return bufptr;
 }
